Update with new URLs
