--- spec/api/shipping_api_spec.rb
@@ -1,83 +1,0 @@
-=begin
-#Patch API V1
-
-#The core API used to integrate with Patch's service
-
-The version of the OpenAPI document: v1
-Contact: engineering@usepatch.com
-Generated by: https://openapi-generator.tech
-OpenAPI Generator version: 5.3.1
-
-=end
-
-require 'spec_helper'
-require 'json'
-
-# Unit tests for Patch::ShippingApi
-# Automatically generated by openapi-generator (https://openapi-generator.tech)
-# Please update as you see appropriate
-describe 'ShippingApi' do
-  before do
-    # run before each test
-    @api_instance = Patch::ShippingApi.new
-  end
-
-  after do
-    # run after each test
-  end
-
-  describe 'test an instance of ShippingApi' do
-    it 'should create an instance of ShippingApi' do
-      expect(@api_instance).to be_instance_of(Patch::ShippingApi)
-    end
-  end
-
-  # unit tests for create_air_shipping_estimate
-  # Creates a GLEC air shipping estimate given freight mass and logistics
-  # Creates a GLEC air shipping estimate for the amount of CO2 to be compensated. An order in the &#x60;draft&#x60; state may be created based on the parameters, linked to the estimate. 
-  # @param create_air_shipping_estimate_request 
-  # @param [Hash] opts the optional parameters
-  # @return [EstimateResponse]
-  describe 'create_air_shipping_estimate test' do
-    it 'should work' do
-      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
-    end
-  end
-
-  # unit tests for create_rail_shipping_estimate
-  # Creates a GLEC rail shipping estimate given freight mass and logistics
-  # Creates a GLEC rail shipping estimate for the amount of CO2 to be compensated. An order in the &#x60;draft&#x60; state may be created based on the parameters, linked to the estimate. 
-  # @param create_rail_shipping_estimate_request 
-  # @param [Hash] opts the optional parameters
-  # @return [EstimateResponse]
-  describe 'create_rail_shipping_estimate test' do
-    it 'should work' do
-      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
-    end
-  end
-
-  # unit tests for create_road_shipping_estimate
-  # Creates a GLEC road shipping estimate given freight mass and logistics
-  # Creates a GLEC road shipping estimate for the amount of CO2 to be compensated. An order in the &#x60;draft&#x60; state may be created based on the parameters, linked to the estimate. 
-  # @param create_road_shipping_estimate_request 
-  # @param [Hash] opts the optional parameters
-  # @return [EstimateResponse]
-  describe 'create_road_shipping_estimate test' do
-    it 'should work' do
-      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
-    end
-  end
-
-  # unit tests for create_sea_shipping_estimate
-  # Creates a GLEC sea shipping estimate given freight mass and logistics
-  # Creates a GLEC sea shipping estimate for the amount of CO2 to be compensated. An order in the &#x60;draft&#x60; state may be created based on the parameters, linked to the estimate. 
-  # @param create_sea_shipping_estimate_request 
-  # @param [Hash] opts the optional parameters
-  # @return [EstimateResponse]
-  describe 'create_sea_shipping_estimate test' do
-    it 'should work' do
-      # assertion here. ref: https://www.relishapp.com/rspec/rspec-expectations/docs/built-in-matchers
-    end
-  end
-
-end
